Separate linux only tests from others
