--- testing/tests/api_tests/linux_only/test_data/winlog_search_data.json
@@ -1,18 +1,0 @@
-{
-    "took": 0,
-    "timed_out": false,
-    "_shards": {
-        "total": 1,
-        "successful": 1,
-        "skipped": 0,
-        "failed": 0
-    },
-    "hits": {
-        "total": {
-            "value": 0,
-            "relation": "eq"
-        },
-        "max_score": null,
-        "hits": []
-    }
-}
