+++ testing/tests/api_tests/winlogbeat/schemas/winlogbeat_search.json
