+++ testing/tests/api_tests/winlogbeat/test_data/mapping_response.json
